core: move json/jsonb functions into turso_core_json
Moves the JSON parser, JSONB codec, path machinery and the scalar
json_*()/jsonb_*() implementations into a turso_core_json crate that only
depends on turso_core_types + turso_core_common. The json_each/json_tree
virtual tables stay in the main crate because they plug into its vtab
machinery; core/json/mod.rs re-exports the moved items so paths do not
change. The dependency is optional and enabled by the existing json
feature.

Tests: cargo check on all four new crates --tests, turso_core with
default/simulator features
