+++ core/crates/io/common.rs
@@ -1,6 +1,6 @@
 pub const ENV_DISABLE_FILE_LOCK: &str = "LIMBO_DISABLE_FILE_LOCK";
 
-#[cfg(test)]
+#[cfg(all(test, feature = "fs"))]
 pub mod tests {
     use crate::{Result, IO};
     use std::process::{Command, Stdio};
