+++ core/crates/json/cache.rs
