+++ core/crates/types/Cargo.toml
@@ -41,3 +41,7 @@ tracing = { workspace = true }
 [target.'cfg(antithesis)'.dependencies]
 antithesis_sdk = { workspace = true, features = ["full"] }
 serde_json = { workspace = true }
+
+[dev-dependencies]
+quickcheck = { version = "1.0", default-features = false }
+quickcheck_macros = { version = "1.0", default-features = false }
